+++ core/src/main/java/com/alibaba/fastjson2/writer/FieldWriterDoubleField.java
@@ -56,7 +56,7 @@ class FieldWriterDoubleField<T>
 
         writeFieldName(jsonWriter);
 
-        double doubleValue = value.doubleValue();
+        double doubleValue = value;
         if (decimalFormat != null) {
             jsonWriter.writeDouble(doubleValue, decimalFormat);
         } else {
@@ -73,7 +73,7 @@ class FieldWriterDoubleField<T>
         if (value == null) {
             jsonWriter.writeNumberNull();
         } else {
-            double doubleValue = value.doubleValue();
+            double doubleValue = value;
             if (decimalFormat != null) {
                 jsonWriter.writeDouble(doubleValue, decimalFormat);
             } else {
