+++ core/src/main/java/com/alibaba/fastjson2/writer/FieldWriterInt8.java
@@ -64,6 +64,6 @@ abstract class FieldWriterInt8<T>
             jsonWriter.writeNumberNull();
             return;
         }
-        jsonWriter.writeInt32(value.byteValue());
+        jsonWriter.writeInt32(value);
     }
 }
