+++ core/src/main/java/com/alibaba/fastjson2/writer/ObjectWriterImplInt8.java
@@ -15,7 +15,7 @@ final class ObjectWriterImplInt8
             return;
         }
 
-        byte byteValue = ((Byte) object).byteValue();
+        byte byteValue = (Byte) object;
         jsonWriter.writeInt8(byteValue);
     }
